Natan changes applied
